Cambios realizados en el ejercicio del triangulo

+++ IDGS903_tema1/Models/Triangulo.cs
@@ -7,11 +7,11 @@ namespace IDGS903_tema1.Models
 {
     public class Triangulo
     {
-        public int X1 { get; set; }
-        public int Y1 { get; set; }
-        public int X2 { get; set; }
-        public int Y2 { get; set; }
-        public int X3 { get; set; }
-        public int Y3 { get; set; }
+        public double X1 { get; set; }
+        public double Y1 { get; set; }
+        public double X2 { get; set; }
+        public double Y2 { get; set; }
+        public double X3 { get; set; }
+        public double Y3 { get; set; }
     }
 }
